--- src/init/consented/prebid.ts
@@ -1,25 +1,0 @@
-import pbjs from 'prebid.js';
-import 'prebid.js/modules/adyoulikeBidAdapter';
-import 'prebid.js/modules/consentManagementTcf';
-import 'prebid.js/modules/consentManagementUsp';
-import 'prebid.js/modules/criteoBidAdapter';
-import 'prebid.js/modules/gridBidAdapter';
-import 'prebid.js/modules/ixBidAdapter';
-import 'prebid.js/modules/ozoneBidAdapter';
-import 'prebid.js/modules/permutiveRtdProvider';
-import 'prebid.js/modules/prebidServerBidAdapter';
-import 'prebid.js/modules/pubmaticBidAdapter';
-import 'prebid.js/modules/sharedIdSystem';
-import 'prebid.js/modules/tripleliftBidAdapter';
-import 'prebid.js/modules/kargoBidAdapter';
-import 'prebid.js/modules/rubiconBidAdapter';
-import 'prebid.js/modules/ttdBidAdapter';
-
-// Guardian specific adapters that we have modified or created
-import '../../lib/header-bidding/prebid/custom-modules/appnexusBidAdapter';
-import '../../lib/header-bidding/prebid/custom-modules/openxBidAdapter';
-import '../../lib/header-bidding/prebid/custom-modules/guAnalyticsAdapter';
-
-pbjs.processQueue();
-
-export { pbjs };
