+++ src/lib/header-bidding/prebid/initialise.spec.ts
@@ -2,7 +2,7 @@ import { type ConsentState } from '@guardian/libs';
 import { pubmatic } from '../../__vendor/pubmatic';
 import { getAdvertById as getAdvertById_ } from '../../dfp/get-advert-by-id';
 import { shouldIncludePermutive } from '../utils';
-import { prebid } from './prebid';
+import { prebid } from './initialise';
 
 const getAdvertById = getAdvertById_ as jest.Mock;
